Move CLI from scripts to entrypoints
This make these easier to test and also means we are using a
non-deprecated pyproject.toml configuration for them.
