+++ src/npg_irods/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Command line interface for npg-irods-python."""
