+++ src/npg_irods/cli/check_checksums.py
@@ -1,7 +1,6 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2023 Genome Research Ltd. All rights reserved.
+# Copyright © 2022, 2023 Genome Research Ltd. All rights reserved.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -23,36 +22,28 @@ import sys
 
 import structlog
 
-from npg_irods.utilities import (
-    check_consent_withdrawn,
-)
-from npg_irods.cli import add_logging_arguments, configure_logging
+from npg_irods.utilities import check_checksums
+from npg_irods.cli.util import add_logging_arguments, configure_logging
 from npg_irods.version import version
 
 description = """
-Reads iRODS data object paths from a file or STDIN, one per line and checks that each
-one is in a state consistent with sample consent being withdrawn.
+Reads iRODS data object paths from a file or STDIN, one per line and performs
+consistency checks on their iRODS checksums and checksum metadata.
 
-The conditions for data objects to be in the correct state for having had consent
-withdrawn are: 
+The conditions for checksums and checksum metadata of a data object to be correct
+are:
 
- - The data object has the correct metadata. Either:
-     - sample_consent = 0 (data managed by the GAPI codebase)
-     - sample_consent_withdrawn = 1 (data managed by the NPG codebase)
+- The data object must have a checksum set for each valid replica.
+- The checksums for all replicas must have the same value.
+- The data object must have one, and only one, checksum AVU in its metadata.
+- The checksum AVU must have the same value as the replica checksums.
 
- - Read permission for any SequenceScape study iRODS groups (named ss_<Study ID>)
-   absent.
+This script will never change any values. To repair checksums, use the --print-fail
+option to print failed paths to STDOUT and pipe them to the desired repair script
+e.g. `repair-checksums`.
 
-N.B. having this consent withdrawn state is only the first step in handling consent
-removal. Subsequent steps, including redaction of data within data objects are not
-handled.
-
-This script will never change any values. To withdraw consent, use the --print-fail
-option to print failed paths to STDOUT and pipe them to the `withdraw-consent` script
-included in this package.
-
-If any of the paths fail their check, the exit code will be non-zero and an error
-message summarising the results will be sent to STDERR.
+If any of the paths fail their checksum check, the exit code will be non-zero and an 
+error message summarising the results will be sent to STDERR.
 """
 
 parser = argparse.ArgumentParser(
@@ -83,8 +74,22 @@ parser.add_argument(
     help="Print to output those paths that fail the check. Defaults to False.",
     action="store_true",
 )
-parser.add_argument("--version", help="Print the version and exit", action="store_true")
 
+parser.add_argument(
+    "-c",
+    "--clients",
+    help="Number of baton clients to use. Defaults to 4.",
+    type=int,
+    default=4,
+)
+parser.add_argument(
+    "-t",
+    "--threads",
+    help="Number of threads to use. Defaults to 4.",
+    type=int,
+    default=4,
+)
+parser.add_argument("--version", help="Print the version and exit", action="store_true")
 
 args = parser.parse_args()
 configure_logging(
@@ -102,11 +107,13 @@ def main():
         print(version())
         exit(0)
 
-    num_processed, num_passed, num_errors = check_consent_withdrawn(
+    num_processed, num_passed, num_errors = check_checksums(
         args.input,
         args.output,
         print_pass=args.print_pass,
         print_fail=args.print_fail,
+        num_clients=args.clients,
+        num_threads=args.threads,
     )
 
     if num_errors:
@@ -124,7 +131,3 @@ def main():
         num_passed=num_passed,
         num_errors=num_errors,
     )
-
-
-if __name__ == "__main__":
-    main()
